requirements/base.txt
Django>=2.0
psycopg2-binary
django-rest-swagger
djangorestframework
django-cors-headers